move business logic into core
